+++ CHANGE_LOG.md
@@ -1,7 +1,13 @@
 ## 1.0.6
 
+#### New Features
+  - added the output of `get_fundamentals()` to the output of `get_tickers()`
+
 #### Clean up
   - `get_historicals()` now returns prices formatted as numeric
+  - `get_fundamentals()` now returns as a data frame instead of a list
+
+
 
 ## 1.0.5